Script Engine Update Phase 1
Dropping array limits, test phase.
Also Implements a under-development (u)int64 key support to the dbmaps.
Made Possible Thanks to Haruna!

+++ src/map/mapreg_sql.c
@@ -20,29 +20,31 @@ struct mapreg_interface mapreg_s;
 
 #define MAPREG_AUTOSAVE_INTERVAL (300*1000)
 
+/* TODO may be worth not to save/store arrays with the SCRIPT_ARRAY_LEN_IDX index */
+
 /// Looks up the value of an integer variable using its uid.
-int mapreg_readreg(int uid) {
-	struct mapreg_save *m = idb_get(mapreg->db, uid);
+int mapreg_readreg(int64 uid) {
+	struct mapreg_save *m = i64db_get(mapreg->db, uid);
 	return m?m->u.i:0;
 }
 
 /// Looks up the value of a string variable using its uid.
-char* mapreg_readregstr(int uid) {
-	struct mapreg_save *m = idb_get(mapreg->str_db, uid);
+char* mapreg_readregstr(int64 uid) {
+	struct mapreg_save *m = i64db_get(mapreg->str_db, uid);
 	return m?m->u.str:NULL;
 }
 
 /// Modifies the value of an integer variable.
-bool mapreg_setreg(int uid, int val) {
+bool mapreg_setreg(int64 uid, int val) {
 	struct mapreg_save *m;
-	int num = (uid & 0x00ffffff);
-	int i   = (uid & 0xff000000) >> 24;
+	int num = (int32)(uid & 0xFFFFFFFF);
+	int i   = (int32)((uid >> 32) & 0xFFFFFFFF);
 	const char* name = script->get_str(num);
 
 	if( val != 0 ) {
-		if( (m = idb_get(mapreg->db,uid)) ) {
+		if( (m = i64db_get(mapreg->db,uid)) ) {
 			m->u.i = val;
-			if(name[1] != '@') {
+			if(name[1] != '@' && i != SCRIPT_ARRAY_LEN_IDX) {// $@ isn't saved; index SCRIPT_ARRAY_LEN_IDX also isn't saved;
 				m->save = true;
 				mapreg->i_dirty = true;
 			}
@@ -53,21 +55,22 @@ bool mapreg_setreg(int uid, int val) {
 			m->uid = uid;
 			m->save = false;
 
-			if(name[1] != '@') {// write new variable to database
+			if(name[1] != '@' && i != SCRIPT_ARRAY_LEN_IDX) {// $@ isn't saved; index SCRIPT_ARRAY_LEN_IDX also isn't saved;
+				// write new variable to database
 				char tmp_str[32*2+1];
 				SQL->EscapeStringLen(map->mysql_handle, tmp_str, name, strnlen(name, 32));
 				if( SQL_ERROR == SQL->Query(map->mysql_handle, "INSERT INTO `%s`(`varname`,`index`,`value`) VALUES ('%s','%d','%d')", mapreg->table, tmp_str, i, val) )
 					Sql_ShowDebug(map->mysql_handle);
 			}
-			idb_put(mapreg->db, uid, m);
+			i64db_put(mapreg->db, uid, m);
 		}
 	} else { // val == 0
-		if( (m = idb_get(mapreg->db,uid)) ) {
+		if( (m = i64db_get(mapreg->db,uid)) ) {
 			ers_free(mapreg->ers, m);
 		}
-		idb_remove(mapreg->db,uid);
+		i64db_remove(mapreg->db,uid);
 
-		if( name[1] != '@' ) {// Remove from database because it is unused.
+		if( name[1] != '@' && i != SCRIPT_ARRAY_LEN_IDX) {// $@ isn't saved; index SCRIPT_ARRAY_LEN_IDX also isn't saved;
 			if( SQL_ERROR == SQL->Query(map->mysql_handle, "DELETE FROM `%s` WHERE `varname`='%s' AND `index`='%d'", mapreg->table, name, i) )
 				Sql_ShowDebug(map->mysql_handle);
 		}
@@ -77,29 +80,29 @@ bool mapreg_setreg(int uid, int val) {
 }
 
 /// Modifies the value of a string variable.
-bool mapreg_setregstr(int uid, const char* str) {
+bool mapreg_setregstr(int64 uid, const char* str) {
 	struct mapreg_save *m;
-	int num = (uid & 0x00ffffff);
-	int i   = (uid & 0xff000000) >> 24;
+	int num = (int)(uid & 0xFFFFFFFF);
+	int i   = (int32)((uid >> 32) & 0xFFFFFFFF);
 	const char* name = script->get_str(num);
 	
 	if( str == NULL || *str == 0 ) {
-		if(name[1] != '@') {
+		if(name[1] != '@' && i != SCRIPT_ARRAY_LEN_IDX) {// $@ isn't saved; index SCRIPT_ARRAY_LEN_IDX also isn't saved;
 			if( SQL_ERROR == SQL->Query(map->mysql_handle, "DELETE FROM `%s` WHERE `varname`='%s' AND `index`='%d'", mapreg->table, name, i) )
 				Sql_ShowDebug(map->mysql_handle);
 		}
-		if( (m = idb_get(mapreg->str_db,uid)) ) {
+		if( (m = i64db_get(mapreg->str_db,uid)) ) {
 			if( m->u.str != NULL )
 				aFree(m->u.str);
 			ers_free(mapreg->ers, m);
 		}
-		idb_remove(mapreg->str_db,uid);
+		i64db_remove(mapreg->str_db,uid);
 	} else {
-		if( (m = idb_get(mapreg->str_db,uid)) ) {
+		if( (m = i64db_get(mapreg->str_db,uid)) ) {
 			if( m->u.str != NULL )
 				aFree(m->u.str);
 			m->u.str = aStrdup(str);
-			if(name[1] != '@') {
+			if(name[1] != '@' && i != SCRIPT_ARRAY_LEN_IDX) {// $@ isn't saved; index SCRIPT_ARRAY_LEN_IDX also isn't saved;
 				mapreg->str_dirty = true;
 				m->save = true;
 			}
@@ -110,7 +113,8 @@ bool mapreg_setregstr(int uid, const char* str) {
 			m->u.str = aStrdup(str);
 			m->save = false;
 			
-			if(name[1] != '@') { //put returned null, so we must insert.
+			if(name[1] != '@' && i != SCRIPT_ARRAY_LEN_IDX) {// $@ isn't saved; index SCRIPT_ARRAY_LEN_IDX also isn't saved;
+				//put returned null, so we must insert.
 				char tmp_str[32*2+1];
 				char tmp_str2[255*2+1];
 				SQL->EscapeStringLen(map->mysql_handle, tmp_str, name, strnlen(name, 32));
@@ -118,7 +122,7 @@ bool mapreg_setregstr(int uid, const char* str) {
 				if( SQL_ERROR == SQL->Query(map->mysql_handle, "INSERT INTO `%s`(`varname`,`index`,`value`) VALUES ('%s','%d','%s')", mapreg->table, tmp_str, i, tmp_str2) )
 					Sql_ShowDebug(map->mysql_handle);
 			}
-			idb_put(mapreg->str_db, uid, m);
+			i64db_put(mapreg->str_db, uid, m);
 		}
 	}
 
@@ -155,28 +159,60 @@ void script_load_mapreg(void) {
 		struct mapreg_save *m = NULL;
 		int s = script->add_str(varname);
 		int i = index;
+		DBMap *mdb = mapreg->db;
 
 		if( varname[length-1] == '$' ) {
-			if( idb_exists(mapreg->str_db, (i<<24)|s) ) {
-				ShowWarning("load_mapreg: duplicate! '%s' => '%s' skipping...\n",varname,value);
-				continue;
-			}
-		} else {
-			if( idb_exists(mapreg->db, (i<<24)|s) ) {
-				ShowWarning("load_mapreg: duplicate! '%s' => '%s' skipping...\n",varname,value);
-				continue;
-			}
+			mdb = mapreg->str_db;
+		}
+		
+		if( i64db_exists(mdb, reference_uid(s,i)) ) {
+			ShowWarning("load_mapreg: duplicate! '%s' (%d) => '%s' skipping...\n",varname,index,value);
+			continue;
 		}
 		
 		m = ers_alloc(mapreg->ers, struct mapreg_save);
-		m->uid = (i<<24)|s;
+		m->uid = reference_uid(s,i);
 		m->save = false;
+		
 		if( varname[length-1] == '$' ) {
 			m->u.str = aStrdup(value);
-			idb_put(mapreg->str_db, m->uid, m);
 		} else {
 			m->u.i = atoi(value);
-			idb_put(mapreg->db, m->uid, m);
+		}
+		
+		i64db_put(mdb, m->uid, m);
+		
+		if( i > 0 ) {
+			if( (m = i64db_get(mdb, reference_uid(s,SCRIPT_ARRAY_LEN_IDX))) ) {
+				if( varname[length-1] == '$' ) {
+					int len;
+					
+					if( (len = atoi(m->u.str)) < i ) {
+						char v[11];
+						
+						aFree(m->u.str);
+						snprintf(v,11,"%d",i+1);
+						m->u.str = aStrdup(v);
+					}
+				} else {
+					if( m->u.i < i )
+						m->u.i = i+1;
+				}
+			} else {
+				m = ers_alloc(mapreg->ers, struct mapreg_save);
+				m->uid = reference_uid(s,SCRIPT_ARRAY_LEN_IDX);
+				m->save = false;
+				
+				if( varname[length-1] == '$' ) {
+					char v[11];
+					
+					snprintf(v,11,"%d",i+1);
+					m->u.str = aStrdup(v);
+				} else {
+					m->u.i = i+1;
+				}
+				i64db_put(mdb, m->uid, m);
+			}
 		}
 	}
 	
@@ -195,8 +231,8 @@ void script_save_mapreg(void) {
 		iter = db_iterator(mapreg->db);
 		for( m = dbi_first(iter); dbi_exists(iter); m = dbi_next(iter) ) {
 			if( m->save ) {
-				int num = (m->uid & 0x00ffffff);
-				int i   = (m->uid & 0xff000000) >> 24;
+				int num = (int32)(m->uid & 0xFFFFFFFF);
+				int i   = (int32)((m->uid >> 32) & 0xFFFFFFFF);
 				const char* name = script->get_str(num);
 
 				if( SQL_ERROR == SQL->Query(map->mysql_handle, "UPDATE `%s` SET `value`='%d' WHERE `varname`='%s' AND `index`='%d' LIMIT 1", mapreg->table, m->u.i, name, i) )
@@ -212,8 +248,8 @@ void script_save_mapreg(void) {
 		iter = db_iterator(mapreg->str_db);
 		for( m = dbi_first(iter); dbi_exists(iter); m = dbi_next(iter) ) {
 			if( m->save ) {
-				int num = (m->uid & 0x00ffffff);
-				int i   = (m->uid & 0xff000000) >> 24;
+				int num = (int32)(m->uid & 0xFFFFFFFF);
+				int i   = (int32)((m->uid >> 32) & 0xFFFFFFFF);
 				const char* name = script->get_str(num);
 				char tmp_str2[2*255+1];
 
@@ -289,8 +325,8 @@ void mapreg_final(void) {
 }
 
 void mapreg_init(void) {
-	mapreg->db = idb_alloc(DB_OPT_BASE);
-	mapreg->str_db = idb_alloc(DB_OPT_BASE);
+	mapreg->db = i64db_alloc(DB_OPT_BASE);
+	mapreg->str_db = i64db_alloc(DB_OPT_BASE);
 	mapreg->ers = ers_new(sizeof(struct mapreg_save), "mapreg_sql.c::mapreg_ers", ERS_OPT_NONE);
 
 	mapreg->load();
